backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
asyncpg==0.31.0
pyjwt==2.8.0
python-dotenv==1.0.0
pytest==8.0.0
httpx==0.26.0
aiosqlite==0.19.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
